Solve issue : In some screens, switching pagination row number does sometimes not refresh correctly the tables (#8004)
